slack_to_fastapi/requirements.txt
python-dotenv
fastapi[standard-no-fastapi-cloud-cli]
slack-bolt
aiohttp
